Discover LegacyVueRouterRedirectCustomPathTest in its own file
PHPUnit only loads one class per *Test.php file by default; moving the
custom path/proxy redirect cases into a matching filename ensures they run.
